refactor(26): extract EvolutionTaskDispatcher from evolution-worker
Extract task dispatch and execution for pain_diagnosis and sleep_reflection
into dedicated EvolutionTaskDispatcher module following Phase 24/25 pattern.
evolution-worker.ts now delegates all dispatch to the new module.
